apache/app/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.2
wheel==0.24.0
redis==2.10.6
mysql-connector-python==8.0.13
